Gets to the character select screen.

+++ Core/IPayload.cs
@@ -5,7 +5,7 @@ namespace Core
 	public interface IPayload
 	{
 		MessageType.ServerToClient Type { get; }
-		ushort Length { get; }
+		int Length { get; }
 		void Marshal(Span<byte> span);
 	}
 }
